% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nep.R
\name{nep}
\alias{nep}
\title{Wrapper function to perform network enhancement and prioritize using MultiNEP or DiSNEP}
\usage{
nep(
  s0,
  E,
  signals,
  signal_type = "pvalue",
  alpha = 0.75,
  beta = 0.75,
  feature_name_list,
  post_process_percent = c(0.95, 0.7, 0.85),
  iter = 30,
  difference = 1e-06,
  return_mat = T,
  lambda_g = 0.05,
  lambda_m = 20,
  model = "multinep",
  parallel = F,
  nworker = 4
)
}
\arguments{
\item{s0}{matrix of the general network}

\item{E}{disease similarity matrix}

\item{signals}{dataframe of initial disease-association signal scores, first column is gene/metabolite names, second column is initial scores}

\item{signal_type}{choose from "pvalue" (default) or "score"}

\item{alpha}{default value is 0.75 to control the degree to use disease omics profiles for enhancement}

\item{beta}{default value is 0.75 to control the degree to use enhanced network to update signals}

\item{feature_name_list}{list for omics feature names, should be length of 2}

\item{post_process_percent}{a vector of thresholds for post-processing enhanced disease-specific matrix. Should be in order of: c(gene network, metabolite network, gene-metabolite network)}

\item{iter}{iteration number, default as 30}

\item{difference}{iteration stop criteria}

\item{return_mat}{if TRUE, return a list for enhanced networks before and after postprocessing}

\item{lambda_g}{value to downweight gene-gene network to gene-metabolite network}

\item{lambda_m}{value to upweight metabolite-metabolite network to gene-metabolite network}

\item{model}{choose from 'multinep' and 'disnep'}

\item{parallel}{whether opt for parallel computing}

\item{nworker}{if opt for parallel computing, how many works to use}
}
\value{
result list with the first element being prioritized signals, the second is a list of enhanced networks if set return_mat as TRUE
}
\description{
Wrapper function to perform network enhancement and prioritize using MultiNEP or DiSNEP
}
\examples{
nep(s0,E,signals,signal_type = 'pvalue',feature_name_list=list(gene_names,metabolite_names),
post_process_percent=c(0.95,0.7,0.85),lambda_g=0.05,lambda_m=20,model = 'multinep')
}
